pyproject.toml
[project]
name = "crypto-alerts-backend"
version = "0.1.0"
description = "Crypto Price Alert System Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.9.1",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "ta-lib>=0.6.7",
    "seaborn>=0.13.2",
    "plotly>=6.3.1",
    "ccxt>=4.0.0",
]
